Refactor tariff services and controllers
- Updated AdminCountryController to return ResponseEntity for createCountry method.
- Removed unused measureId, preferenceId, suspensionId, and tid setters in AdminMeasureController, AdminPreferenceController, AdminSuspensionController, and AdminTransactionController.
- Changed measureId, preferenceId, suspensionId, and tid types from Long to Integer in respective DTOs and models.
- Added new admin services for Country, Measure, Preference, Product, Transaction, and User with CRUD operations.
- Removed old service implementations for Country, Product, and TariffRate.
- Updated Product model to set length for hs6Code column.
- Adjusted User table name to escape reserved keyword.

+++ tariff/src/main/java/com/verbosegarbonzo/tariff/model/Measure.java
@@ -16,7 +16,7 @@ public class Measure {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "measure_id")
-    private Long measureId;
+    private Integer measureId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "importer_code", referencedColumnName = "country_code", nullable = false)
